Debugger: Fix a warning
Followup to 01006580c10a4b9f8be689af61e90cd0f9134688

+++ src/plugins/debugger/debuggerprotocol.cpp
@@ -267,7 +267,7 @@ static QByteArray ind(int indent)
 
 void GdbMi::dumpChildren(QByteArray * str, bool multiline, int indent) const
 {
-    for (size_t i = 0; i < m_children.size(); ++i) {
+    for (int i = 0; i < m_children.size(); ++i) {
         if (i != 0) {
             *str += ',';
             if (multiline)
